requirements.txt
acora==2.2
AdvancedHTTPServer==2.2.0
aioredis==1.3.1
asgiref==3.2.10
asn1crypto==1.4.0
async-timeout==3.0.1
asysocks==0.0.2
attrs==19.3.0
autobahn==17.10.1
Automat==20.2.0
cffi==1.14.1
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==2.8
Cython==0.29.21
daphne==3.0.1
dj-database-url==0.5.0
Django==3.1
django-channels==0.7.0
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-debug-toolbar==2.1
django-excel-response2==2.0.8
django-heroku==0.3.1
django-js-asset==1.2.2
django-registration-redux==2.8
django-restframework==0.0.1
django-six==1.0.4
djangorestframework==3.12.2
gunicorn==20.0.4
hiredis==1.1.0
httptools==0.1.1
hyperlink==20.0.1
Pillow==7.2.0
psycopg2==2.8.5         
psycopg2-binary==2.8.6    
idna==2.10
incremental==16.10.1
msgpack==1.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.20
PyHamcrest==1.9.0
PyOpenGL==3.1.5
pyparsing==2.4.6
pytz==2019.3
redis==3.3.11
requests==2.23.0
service-identity==18.1.0
six==1.13.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
sqlalchemy-schemadisplay==1.3
sqlparse==0.3.0
txaio==20.4.1
urllib3==1.25.9
uvicorn==0.13.3
uvloop==0.15.0
whitenoise==5.2.0